+++ storage/src/status.cc
@@ -51,7 +51,7 @@ status::status()
  * @param Unknown
  */
 status::status(timestamp const& ctime,
-               uint32_t index_id,
+               uint64_t index_id,
                uint32_t interval,
                bool is_for_rebuild,
                timestamp const& rrd_len,
